requirements.txt
twilio==8.10.0
openai==1.12.0
elevenlabs==0.2.27
google-cloud-speech==2.23.0
google-cloud-texttospeech==2.14.1
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
